pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "everyone-pm"
version = "0.2.6"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.81.13",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
    "typer>=0.24.0",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.12.3",
    "pymupdf>=1.26.0",
    "questionary>=2.0.1",
]

[project.scripts]
eopm = "eopm.main:main"

[tool.setuptools]
packages = ["eopm", "eopm.core", "eopm.llm", "eopm.ui"]
